[Parser] In Parser::parseBraceItemList, when we don't find a right brace recover by using the PreviousLoc.
This ensures that the brace range will include all contained items.
The previous scheme of choosing the end location of the last item was not accurate because for pattern bindings decls
we also push the var decls as items, but they don't have the full source range of the pattern binding decl.

Fixes rdar://15190792.

Swift SVN r9293