Step #2 of the load promotion rewrite: change how assigns are handled.
Previously, the definite init pass would check to see if there was a
load available to optimize out the load of a non-trivial assign.  Now
we just unconditionally lower an assign to a load/copy/store sequence,
and lead later load elimination zap the load if it is redundant.

This allows us to remove the old mechanics for doing reasoning about
load elimination, including all the "AccessPath" stuff.


Swift SVN r9340