Use disjunctions to decide between tuple-to-tuple and scalar-to-tuple.
Replace some really awful heuristics used to guess whether the
tuple-to-tuple or scalar-to-tuple conversion should be applied with a
disjunction constraint, where we favor the tuple-to-tuple conversion
when it succeeds. This fixes <rdar://problem/12985801> and the
majority of <rdar://problem/11293232> (all test cases in both work).

There is still some cleanup to do, and we still get some unfortunate
deductions such as

  (swift) [(1, 2)]
  // r0 : Int[] = [1, 2]

that will need to be addressed.


Swift SVN r9407