Synchronize a generic function type's requirements with archetype requirements.
The list of requirements within a generic function type is
functionally identical to the list of conformances in the list of "all
archetypes" that a given function carries. Synchronize these lists so
they have identical requirements in the same order, which allows us to
substitute the former for the latter.


Swift SVN r9472