SILGen: Implement NominalTypePatterns for stored properties.
If we see nominal type patterns in a switch column, grab all of the needed properties for all of the patterns when destructuring so we only need to do it once. For now, only handle stored properties. John's in the middle of reinventing the world for SIL functions and I don't want to create merge problems.

There's an issue handling the temporary allocation for address-only properties that I don't have time to address right now but will fix soon.

Swift SVN r9520