SILGen: Use "constructor prefix" heuristic to optimize switch decision trees.
According to Maranget this heuristic gave the best results over a large corpus of Ocaml code. Luckily, it's really easy to implement too—just count through each column until we hit a wildcard, and specialize on the column with the most non-wildcard pattern nodes before a wildcard.

Swift SVN r9533