SILGenPattern: Let 'is T' and 'U(x:, y:)' patterns coexist.
When destructuring a NominalTypePattern, keep the aggregate around as an extra column of the destructured matrix so we can still match non-NominalTypePatterns against it. This handles 'is T' patterns on the same type, and should also handle other fancy high-level patterns we might add in the future.

Swift SVN r9538