Some belated code review of CapturePromotion.  Much of it is minor typographical/style changes and use of simpler/strength reduced apis (e.g. empty() instead of size() != 0).
Stephen, I added a couple of TODOs for simple ways that the
algorithm could be made better, and added a few FIXMEs for
higher priority things that should be taken care of.  The
biggest issue is that we need to find a better datastructure
for the reachability queries.


Swift SVN r9546