Teach GenericTypeResolver to produce the type of a DeclContext.
When we're checking the signature of a generic function, we want
references to elements within an enclosing DeclContext (such as an
unspecialized reference to an enclosing generic class or a reference
to a typealias of a generic class) to involve generic parameters
rather than archetypes. Elsewhere, we want the
archetypes. GenericTypeResolver is responsible for handling this
distinction.


Swift SVN r9563