[CCP] Rework CCP to use an inverted algorithm.
Instead of adding all the instructions into the WorkList to begin with,
add only the constants and try to fold their users. Each time a user can
be folded, add it to the WorkList. This keeps the WorkList smaller.

The inverted algorithm also allows to simplify the handling of struct_extract
and tuple_extract.

(as suggested by Chris in comments on r9545)

Swift SVN r9598