@objc inference: fix inference for getters and setters
While computing ObjC-ness of a getter, we call ->isObjC() on the parent
property potentially before it was type checked.  The source contains a comment
saying that one should refrain from doing so, but the code does anyway.
I don't think that it is appropriate to follow the advice in the comment and
check the raw form of attributes directly.

Instead, just validate the property declaration before calling ->isObjC().


Swift SVN r9841