Sema: Resolve ExprPatterns using swift.~= if ~= doesn't exist in the current context.
This is a bit of a hack. Clang module contexts currently don't normally see the standard library, so a RawRepresentable conformance synthesized into a Clang module fails to type-check because its 'switch' statements can't resolve '~='. The better answer here might be to import the stdlib into Clang modules, but Jordan and I aren't sure if that has unforeseen consequences.

Swift SVN r9987