[utils] buildbot-script.sh: add command-line option system
These changes create a viable way for people to build and test Swift
locally with reasonable assurance that the buildbot is testing the
same thing they are.  A number of other changes are planned, but these
are the basics and we need to make sure the 'bot still works before
proceeding.

Along the way, I had Joe change the way the buildbot invokes this
script so that $WORKSPACE is now passed on the command line.  Passing
flags in environment variables is inherently fragile, so the script no
longer uses that configuration method.

Swift SVN r10357