Rework the applicable-function constraint to handle its own argument conversion.
Rather than introducing a new type variable for every
applicable-function constraint's input type, place the actual argument
type into that position instead. Aside from eliminating a useless type
variable, this is a step toward further improvements with function
application.

Note that there are a few issues uncovered or introduced by this
patch. It makes <rdar://problem/15172575> slightly worse, which caused
the change in GetOpt where we were getting away with overloading on T
vs. T?. It also makes our recovery for code-completion purposes worse
in one of the tests, but this is a general problem
(<rdar://problem/15470973>).


Swift SVN r10462