Allow construction on arbitrary metatype values.
Allow ApplicableFunction constraints to be solved by metatypes, which
turns into a construction constraint. This allows the expression used
as the 'function' in the application to be an arbitrary expression
that produces a metatype. Previously, we could only handle very
limited forms of type references.

We'll need to introduce some restrictions for class types here; this
will come when we implement virtual initializers.


Swift SVN r10472