Diagnose partial applications of (non-class-bound) generic methods.
These have the same semantic issue as partial applications of structs or enums because the generic parameter or existential may be bound to a value type, so let's ban them for consistency till we get our story straight here. While we're here, improve the error message to be a bit more contextually aware.

Swift SVN r10603