ingloriously fix the serious DI problem: <rdar://problem/15511392> DI cannot promote loads until the initialization semantics of an entire function is known
This splits load (+destroy_addr) promotion out to a separate pass that
happens after the DI properties of an entire function are known, resolving
the problem.

The upshot is that with this fix, DI is now feature complete (and correct!) for 
all known local and global variable situations (just leaving initializers to do).

The downshot is that some cleanup is required, because this is a purely mechanical
change to the previous algorithms.



Swift SVN r10702