[sil-combine] Change replaceInstUsesWith so that it replaces all results instead of just the 0 result.
I am preparing an additional patch after this that gives the user the option of
just replacing the ith result of the instruction with the jth result of the
value type. Only the users of the ith result of the instruction in that method
will be added to the worklist.

Swift SVN r10755